% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_packages2.R
\name{current_packages2}
\alias{current_packages2}
\title{\itemize{
\item New behaviour:
\itemize{
\item Should check a global option
\item Then any \code{library()} calls
\item Then the \code{DESCRIPTION} file
\item Then use search path - but you should have to opt in to this because
it can easily go wrong.
\item Should be much chattier
}
}}
\usage{
current_packages2(use = c("global_option", "library_calls", "package_desc"))
}
\description{
\itemize{
\item New behaviour:
\itemize{
\item Should check a global option
\item Then any \code{library()} calls
\item Then the \code{DESCRIPTION} file
\item Then use search path - but you should have to opt in to this because
it can easily go wrong.
\item Should be much chattier
}
}
}
