% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_packages2.R
\name{guess_pkgs}
\alias{guess_pkgs}
\title{\itemize{
\item New behaviour:
\itemize{
\item Should check a global option
\item Then any \code{library()} calls in selected code.
\item Then the \code{DESCRIPTION} file if present
\item Then use search path - but you should have to opt in to this because
it can easily go wrong.
\item Should be much chattier
}
}}
\usage{
guess_pkgs(
  code = NULL,
  order = c("global_option", "package_desc", "library_calls")
)
}
\description{
\itemize{
\item New behaviour:
\itemize{
\item Should check a global option
\item Then any \code{library()} calls in selected code.
\item Then the \code{DESCRIPTION} file if present
\item Then use search path - but you should have to opt in to this because
it can easily go wrong.
\item Should be much chattier
}
}
}
