% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positron_extension.R
\name{ic}
\alias{ic}
\title{Format File with Error Handling}
\usage{
ic(path)
}
\arguments{
\item{path}{A character string specifying the path to the file with
code to be formatted.}
}
\value{
None
}
\description{
This function reads the content of a file at the specified path,
makes R function calls explicit using the \code{pedant::add_double_colons} function,
and then writes the updated content back to the file.
If the file does not exist, it writes 'Error_002' to the specified path.
If an error occurs during the reading or processing of the file,
it writes 'Error_001' to the specified path.
}
