% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_dcs.R
\name{insert_dcs}
\alias{insert_dcs}
\title{Title}
\usage{
insert_dcs(code, use_pkgs = NULL, ignore_funs = NULL, src = NULL)
}
\arguments{
\item{code}{}

\item{use_pkgs}{}

\item{ignore_funs}{}

\item{src}{}
}
\description{
Title
}
