% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedant-package.R
\docType{package}
\name{pedant-package}
\alias{pedant}
\alias{pedant-package}
\title{pedant: Make Function Calls Explicit}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Transform code from using implicit function calls to explicit using the double-colon operator.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wurli/pedant}
  \item Report bugs at \url{https://github.com/wurli/pedant/issues}
}

}
\keyword{internal}
