% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_packages2.R
\name{pkgs_from_opt}
\alias{pkgs_from_opt}
\alias{pkgs_from_library_calls}
\alias{pkgs_from_desc}
\alias{pkgs_from_search_path}
\title{Get packages to use when inserting \code{::}}
\usage{
pkgs_from_opt(opt = "pedant_packages")

pkgs_from_library_calls(code, funs = c("library", "require"))

pkgs_from_desc(dir = ".", types = "Imports")

pkgs_from_search_path(search_path = search())
}
\arguments{
\item{opt}{The name of the global option to check}

\item{code}{A character string containing code to search}

\item{funs}{Functions which indicate that a package is available}

\item{dir}{A directory in which to search a \code{DESCRIPTION} and \code{NAMESPACE} files}

\item{types}{A subset of \code{c("Imports", "Depends", "Suggests", "Enhances", "LinkingTo")}}

\item{search_path}{A character vector as returned by \code{search()}}
}
\value{
A character vector of package names or \code{NULL}
}
\description{
\itemize{
\item \code{pkgs_from_opt()} checks a global option
\item \code{pkgs_from_library_calls()} checks code for calls to \code{library()}/\code{require()}
\item \code{pgks_from_desc()} gets the packages listed in the current
project's \code{DESCRIPTION} file, minus any packages which are noted in the
\code{NAMESPACE} file as being imported in their entirety
\item \code{pkgs_from_search_path()} gets the packages which are currently loaded
}
}
\examples{
code <- "library(dplyr)
library(`'tidyr'`)
library(\"ggplot2\")
require('`another_pkg`')
do |> some() |> stuff"

pkgs_from_library_calls(code)
}
