% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.pedant_style_info}
\alias{print.pedant_style_info}
\title{Print information about the last restyle}
\usage{
\method{print}{pedant_style_info}(x, file = NULL, token = NULL, ...)
}
\arguments{
\item{x}{}

\item{file}{}

\item{...}{}
}
\description{
Print information about the last restyle
}
