% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_packages.R
\name{is_dev_context}
\alias{is_dev_context}
\alias{imported_functions}
\alias{current_packages}
\title{Get packages from the current context}
\usage{
is_dev_context(dir = ".")

imported_functions(dir = ".")

current_packages(
  dir = ".",
  base_packages = getOption("defaultPackages"),
  include_types = "Imports"
)
}
\arguments{
\item{dir}{The current working directory}

\item{base_packages}{Default packages to include}

\item{include_types}{The types of package imports to return if the current
context is package development. Should be a subset of
\code{c("Imports", "Depends", "Suggests", "Enhances", "LinkingTo")}}
}
\description{
These functions find the packages/functions to use when running
\code{add_double_colons()}.
}
\details{
\itemize{
\item \code{current_packages()} first checks if the current context is package
development. If it is, then it returns the packages which are listed in the
package \code{DESCRIPTION} as dependencies, but will not return any packages also
listed as imports in the package \code{NAMESPACE}. If the current context is not
package development, the currently attached packages (as given by \code{search()})
are used. Note that if \code{{pkgload}} is not installed then the latter option is
always used.
\item \code{imported_functions()} looks for a package \code{NAMESPACE} file and returns the
names of all imported functions. If a \code{NAMESPACE} file is not found, or if
\code{{pkgload}} is not loaded, \code{NULL} is returned.
}
}
